PRSD-546: Rename and move additional classes
