+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/config/CustomSecurityConfig.kt
@@ -1,10 +1,5 @@
 package uk.gov.communities.prsdb.webapp.config
 
-import jakarta.servlet.Filter
-import jakarta.servlet.FilterChain
-import jakarta.servlet.ServletRequest
-import jakarta.servlet.ServletResponse
-import jakarta.servlet.http.HttpServletRequest
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.core.annotation.Order
@@ -12,16 +7,12 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserRequest
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserService
 import org.springframework.security.oauth2.client.oidc.web.logout.OidcClientInitiatedLogoutSuccessHandler
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserService
-import org.springframework.security.oauth2.client.web.DefaultOAuth2AuthorizationRequestResolver
-import org.springframework.security.oauth2.client.web.OAuth2AuthorizationRequestResolver
-import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest
 import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser
 import org.springframework.security.oauth2.core.oidc.user.OidcUser
 import org.springframework.security.web.SecurityFilterChain
@@ -42,7 +33,6 @@ class CustomSecurityConfig(
             .securityMatchers { config ->
                 config.requestMatchers(
                     AntPathRequestMatcher("/local-authority/**"),
-                    AntPathRequestMatcher("/login/oauth2/code/one-login-id"),
                 )
             }.authorizeHttpRequests { requests ->
                 requests
@@ -53,16 +43,20 @@ class CustomSecurityConfig(
             }.oauth2Login { config ->
                 config.loginPage("/oauth2/authorization/one-login-id")
                 config.authorizationEndpoint { customizer ->
-                    customizer.authorizationRequestResolver(MyRequestResolver(clientRegistrationRepository))
+                    customizer.authorizationRequestResolver(
+                        AdditionalParameterAddingRequestResolver(clientRegistrationRepository, "vtr", "Cl.Cm.P2"),
+                    )
                 }
             }.logout { logout ->
                 logout.logoutSuccessHandler(oidcLogoutSuccessHandler())
             }.csrf { requests ->
                 requests.ignoringRequestMatchers("/local/**")
-            }.addFilterAfter(MyFilter(), SecurityContextHolderFilter::class.java)
+            }.addFilterAfter(OauthTokenSecondaryValidatingFilter(::isOauthTokenValid), SecurityContextHolderFilter::class.java)
 
         return http.build()
     }
+
+    fun isOauthTokenValid(token: OAuth2AuthenticationToken?) = token?.principal?.attributes?.get("vot") == "Cl.Cm.P2"
 
     @Bean
     @Order(2)
@@ -122,44 +116,5 @@ class CustomSecurityConfig(
         oidcLogoutSuccessHandler.setPostLogoutRedirectUri("{baseUrl}/signout")
         oidcLogoutSuccessHandler.setDefaultTargetUrl("/signout")
         return oidcLogoutSuccessHandler
-    }
-}
-
-class MyFilter : Filter {
-    override fun doFilter(
-        request: ServletRequest?,
-        response: ServletResponse?,
-        chain: FilterChain?,
-    ) {
-        val securityContext = SecurityContextHolder.getContext()
-        val auth = securityContext.authentication
-        if (auth != null) {
-            if (!(auth is OAuth2AuthenticationToken && auth.principal.attributes["vot"] == "Cl.Cm.P2")) {
-                SecurityContextHolder.setContext(SecurityContextHolder.createEmptyContext())
-            }
-        }
-
-        chain!!.doFilter(request, response)
-    }
-}
-
-class MyRequestResolver(
-    repo: ClientRegistrationRepository,
-) : OAuth2AuthorizationRequestResolver {
-    private val innerResolver = DefaultOAuth2AuthorizationRequestResolver(repo, "/oauth2/authorization")
-
-    override fun resolve(request: HttpServletRequest?): OAuth2AuthorizationRequest? {
-        val authRequest = innerResolver.resolve(request) ?: return null
-        val additionalParams = LinkedHashMap(authRequest.additionalParameters).apply { put("vtr", "ClCm.P2") }
-        return OAuth2AuthorizationRequest.from(authRequest).additionalParameters(additionalParams).build()
-    }
-
-    override fun resolve(
-        request: HttpServletRequest?,
-        clientRegistrationId: String?,
-    ): OAuth2AuthorizationRequest? {
-        val authRequest = innerResolver.resolve(request, clientRegistrationId) ?: return null
-        val additionalParams = LinkedHashMap(authRequest.additionalParameters).apply { put("vtr", "ClCm.P2") }
-        return OAuth2AuthorizationRequest.from(authRequest).additionalParameters(additionalParams).build()
     }
 }
